third_party/WebKit/LayoutTests/transitions/transitions-parsing.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<div>
  <div style="width:500px;height:500px"><div id="test">hello</div></div>
</div>
<script>
'use strict';
// Test the parsing and the computed style values of the transitions properties.

var e = document.getElementById('test');
var style = e.style;
var computedStyle = window.getComputedStyle(e, null);

// This function checks the return value of computedStyle.transition and verifies WebKit can parse it.
function checkTransitionShorthandValue() {
    var before = computedStyle.getPropertyValue('transition');
    e.style.transition = '';
    e.style.transition = before;
    return (computedStyle.getPropertyValue('transition') == before);
}

test(() => {
  // Initial test.
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "none";
  assert_not_equals(Object.keys(style).indexOf('transitionProperty'), -1);
  assert_equals(Object.keys(style).indexOf('webkitTransitionProperty'), -1);
  assert_equals(style.transitionProperty, 'none');
  assert_equals(computedStyle.transitionProperty, 'none');
  assert_equals(style.webkitTransitionProperty, 'none');
  assert_equals(computedStyle.webkitTransitionProperty, 'none');

  style.transitionProperty = "all";
  assert_equals(style.transitionProperty, 'all');
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(style.webkitTransitionProperty, 'all');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "all, all";
  assert_equals(style.transitionProperty, 'all, all');
  assert_equals(computedStyle.transitionProperty, 'all, all');
  assert_equals(style.webkitTransitionProperty, 'all, all');
  assert_equals(computedStyle.webkitTransitionProperty, 'all, all');

  style.transitionProperty = "-webkit-transform-origin";
  assert_equals(style.transitionProperty, '-webkit-transform-origin');
  assert_equals(computedStyle.transitionProperty, '-webkit-transform-origin');

  style.transitionProperty = "background-position";
  assert_equals(style.transitionProperty, 'background-position');
  assert_equals(computedStyle.transitionProperty, 'background-position');
  assert_equals(style.webkitTransitionProperty, 'background-position');
  assert_equals(computedStyle.webkitTransitionProperty, 'background-position');

  style.transitionProperty = "background-position, font-size";
  assert_equals(style.transitionProperty, 'background-position, font-size');
  assert_equals(computedStyle.transitionProperty, 'background-position, font-size');
  assert_equals(style.webkitTransitionProperty, 'background-position, font-size');
  assert_equals(computedStyle.webkitTransitionProperty, 'background-position, font-size');

  style.transitionProperty = "background-position, font-size, color";
  assert_equals(style.transitionProperty, 'background-position, font-size, color');
  assert_equals(computedStyle.transitionProperty, 'background-position, font-size, color');
  assert_equals(style.webkitTransitionProperty, 'background-position, font-size, color');
  assert_equals(computedStyle.webkitTransitionProperty, 'background-position, font-size, color');

  style.transitionProperty = "all, font-size, color";
  assert_equals(style.transitionProperty, 'all, font-size, color');
  assert_equals(computedStyle.transitionProperty, 'all, font-size, color');
  assert_equals(style.webkitTransitionProperty, 'all, font-size, color');
  assert_equals(computedStyle.webkitTransitionProperty, 'all, font-size, color');

  style.transitionProperty = "font-size, color, all";
  assert_equals(style.transitionProperty, 'font-size, color, all');
  assert_equals(computedStyle.transitionProperty, 'font-size, color, all');
  assert_equals(style.webkitTransitionProperty, 'font-size, color, all');
  assert_equals(computedStyle.webkitTransitionProperty, 'font-size, color, all');

  style.transitionProperty = "font-size, all, color";
  assert_equals(style.transitionProperty, 'font-size, all, color');
  assert_equals(computedStyle.transitionProperty, 'font-size, all, color');
  assert_equals(style.webkitTransitionProperty, 'font-size, all, color');
  assert_equals(computedStyle.webkitTransitionProperty, 'font-size, all, color');
}, "Valid transition-property values.");

test(() => {

  style.transitionProperty = "solid, font-size";
  assert_equals(style.transitionProperty, 'solid, font-size');
  assert_equals(computedStyle.transitionProperty, 'solid, font-size');
  assert_equals(style.webkitTransitionProperty, 'solid, font-size');
  assert_equals(computedStyle.webkitTransitionProperty, 'solid, font-size');

  style.transitionProperty = "solid, left";
  assert_equals(style.transitionProperty, 'solid, left');
  assert_equals(computedStyle.transitionProperty, 'solid, left');
  assert_equals(style.webkitTransitionProperty, 'solid, left');
  assert_equals(computedStyle.webkitTransitionProperty, 'solid, left');

  style.transitionProperty = "solid";
  assert_equals(style.transitionProperty, 'solid');
  assert_equals(computedStyle.transitionProperty, 'solid');
  assert_equals(style.webkitTransitionProperty, 'solid');
  assert_equals(computedStyle.webkitTransitionProperty, 'solid');

  style.transitionProperty = "default, top";
  assert_equals(style.transitionProperty, 'default, top');
  assert_equals(computedStyle.transitionProperty, 'default, top');
  assert_equals(style.webkitTransitionProperty, 'default, top');
  assert_equals(computedStyle.webkitTransitionProperty, 'default, top');
}, "Unknown transition-property values.");

test(() => {
  style.transitionProperty = "";

  style.transitionProperty = "20px";
  assert_equals(style.transitionProperty, '');
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(style.webkitTransitionProperty, '');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "0";
  assert_equals(style.transitionProperty, '');
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(style.webkitTransitionProperty, '');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "'font-size'";
  assert_equals(style.transitionProperty, '');
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(style.webkitTransitionProperty, '');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "all none";
  assert_equals(style.transitionProperty, '');
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(style.webkitTransitionProperty, '');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "opacity width";
  assert_equals(style.transitionProperty, '');
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(style.webkitTransitionProperty, '');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "all, none";
  assert_equals(style.transitionProperty, '');
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(style.webkitTransitionProperty, '');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "none, none";
  assert_equals(style.transitionProperty, '');
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(style.webkitTransitionProperty, '');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "none, all";
  assert_equals(style.transitionProperty, '');
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(style.webkitTransitionProperty, '');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "width, none";
  assert_equals(style.transitionProperty, '');
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(style.webkitTransitionProperty, '');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "none, width";
  assert_equals(style.transitionProperty, '');
  assert_equals(computedStyle.transitionProperty, 'all');
  assert_equals(style.webkitTransitionProperty, '');
  assert_equals(computedStyle.webkitTransitionProperty, 'all');

  style.transitionProperty = "initial, width";
  assert_equals(style.transitionProperty, '');
  style.transitionProperty = "width, inherit";
  assert_equals(style.transitionProperty, '');
  style.transitionProperty = "left, unset";
  assert_equals(style.transitionProperty, '');

  style.transitionProperty = "";
}, "Invalid transition-property values.");

test(() => {
  // Initial test.
  assert_equals(computedStyle.transitionDuration, '0s');
  assert_equals(computedStyle.webkitTransitionDuration, '0s');

  style.transitionDuration = "0s";
  assert_not_equals(Object.keys(style).indexOf('transitionDuration'), -1);
  assert_equals(Object.keys(style).indexOf('webkitTransitionDuration'), -1);
  assert_equals(style.transitionDuration, '0s');
  assert_equals(computedStyle.transitionDuration, '0s');
  assert_equals(style.webkitTransitionDuration, '0s');
  assert_equals(computedStyle.webkitTransitionDuration, '0s');

  style.transitionDuration = "5s";
  assert_equals(style.transitionDuration, '5s');
  assert_equals(computedStyle.transitionDuration, '5s');
  assert_equals(style.webkitTransitionDuration, '5s');
  assert_equals(computedStyle.webkitTransitionDuration, '5s');

  style.transitionDuration = "10ms";
  assert_equals(style.transitionDuration, '10ms');
  assert_equals(computedStyle.transitionDuration, '0.01s');
  assert_equals(style.webkitTransitionDuration, '10ms');
  assert_equals(computedStyle.webkitTransitionDuration, '0.01s');

  style.transitionProperty = "opacity, width";

  style.transitionDuration = "10ms, 20s";
  assert_equals(style.transitionDuration, '10ms, 20s');
  assert_equals(computedStyle.transitionDuration, '0.01s, 20s');
  assert_equals(style.webkitTransitionDuration, '10ms, 20s');
  assert_equals(computedStyle.webkitTransitionDuration, '0.01s, 20s');
}, "Valid transition-duration values.");

test(() => {
  style.transitionProperty = "opacity";
  style.transitionDuration = "";

  // Negative values are invalid.
  style.transitionDuration = "-10ms";
  assert_equals(style.transitionDuration, '');
  assert_equals(computedStyle.transitionDuration, '0s');
  assert_equals(style.webkitTransitionDuration, '');
  assert_equals(computedStyle.webkitTransitionDuration, '0s');

  style.transitionDuration = "'5ms'";
  assert_equals(style.transitionDuration, '');
  assert_equals(computedStyle.transitionDuration, '0s');
  assert_equals(style.webkitTransitionDuration, '');
  assert_equals(computedStyle.webkitTransitionDuration, '0s');

  style.transitionDuration = "30px";
  assert_equals(style.transitionDuration, '');
  assert_equals(computedStyle.transitionDuration, '0s');
  assert_equals(style.webkitTransitionDuration, '');
  assert_equals(computedStyle.webkitTransitionDuration, '0s');

  style.transitionDuration = "solid";
  assert_equals(style.transitionDuration, '');
  assert_equals(computedStyle.transitionDuration, '0s');
  assert_equals(style.webkitTransitionDuration, '');
  assert_equals(computedStyle.webkitTransitionDuration, '0s');

  style.transitionDuration = "20";
  assert_equals(style.transitionDuration, '');
  assert_equals(computedStyle.transitionDuration, '0s');
  assert_equals(style.webkitTransitionDuration, '');
  assert_equals(computedStyle.webkitTransitionDuration, '0s');

  style.transitionDuration = "20%";
  assert_equals(style.transitionDuration, '');
  assert_equals(computedStyle.transitionDuration, '0s');
  assert_equals(style.webkitTransitionDuration, '');
  assert_equals(computedStyle.webkitTransitionDuration, '0s');

  style.transitionDuration = "0s, 20px";
  assert_equals(style.transitionDuration, '');
  assert_equals(computedStyle.transitionDuration, '0s');
  assert_equals(style.webkitTransitionDuration, '');
  assert_equals(computedStyle.webkitTransitionDuration, '0s');
}, "Invalid transition-duration values.");

test(() => {
  // Initial test.
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "linear";
  assert_not_equals(Object.keys(style).indexOf('transitionTimingFunction'), -1);
  assert_equals(Object.keys(style).indexOf('webkitTransitionTimingFunction'), -1);
  assert_equals(style.transitionTimingFunction, 'linear');
  assert_equals(computedStyle.transitionTimingFunction, 'linear');
  assert_equals(style.webkitTransitionTimingFunction, 'linear');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'linear');

  style.transitionTimingFunction = "ease";
  assert_equals(style.transitionTimingFunction, 'ease');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, 'ease');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "ease-in";
  assert_equals(style.transitionTimingFunction, 'ease-in');
  assert_equals(computedStyle.transitionTimingFunction, 'ease-in');
  assert_equals(style.webkitTransitionTimingFunction, 'ease-in');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease-in');

  style.transitionTimingFunction = "ease-out";
  assert_equals(style.transitionTimingFunction, 'ease-out');
  assert_equals(computedStyle.transitionTimingFunction, 'ease-out');
  assert_equals(style.webkitTransitionTimingFunction, 'ease-out');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease-out');

  style.transitionTimingFunction = "ease-in-out";
  assert_equals(style.transitionTimingFunction, 'ease-in-out');
  assert_equals(computedStyle.transitionTimingFunction, 'ease-in-out');
  assert_equals(style.webkitTransitionTimingFunction, 'ease-in-out');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease-in-out');

  style.transitionTimingFunction = "cubic-bezier(0.25, 0.1, 0.25, 1)";
  assert_equals(style.transitionTimingFunction, 'cubic-bezier(0.25, 0.1, 0.25, 1)');
  assert_equals(computedStyle.transitionTimingFunction, 'cubic-bezier(0.25, 0.1, 0.25, 1)');
  assert_equals(style.webkitTransitionTimingFunction, 'cubic-bezier(0.25, 0.1, 0.25, 1)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'cubic-bezier(0.25, 0.1, 0.25, 1)');

  style.transitionTimingFunction = "cubic-bezier(0.1, 0, 0.23, 0.4)";
  assert_equals(style.transitionTimingFunction, 'cubic-bezier(0.1, 0, 0.23, 0.4)');
  assert_equals(computedStyle.transitionTimingFunction, 'cubic-bezier(0.1, 0, 0.23, 0.4)');
  assert_equals(style.webkitTransitionTimingFunction, 'cubic-bezier(0.1, 0, 0.23, 0.4)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'cubic-bezier(0.1, 0, 0.23, 0.4)');

  style.transitionTimingFunction = "cubic-bezier(0.1, 0, 0.23, 3)";
  assert_equals(style.transitionTimingFunction, 'cubic-bezier(0.1, 0, 0.23, 3)');
  assert_equals(computedStyle.transitionTimingFunction, 'cubic-bezier(0.1, 0, 0.23, 3)');
  assert_equals(style.webkitTransitionTimingFunction, 'cubic-bezier(0.1, 0, 0.23, 3)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'cubic-bezier(0.1, 0, 0.23, 3)');

  style.transitionTimingFunction = "cubic-bezier(0.1, 5, 0.23, 3)";
  assert_equals(style.transitionTimingFunction, 'cubic-bezier(0.1, 5, 0.23, 3)');
  assert_equals(computedStyle.transitionTimingFunction, 'cubic-bezier(0.1, 5, 0.23, 3)');
  assert_equals(style.webkitTransitionTimingFunction, 'cubic-bezier(0.1, 5, 0.23, 3)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'cubic-bezier(0.1, 5, 0.23, 3)');

  style.transitionTimingFunction = "cubic-bezier(0.1, 5, 0.23, 0)";
  assert_equals(style.transitionTimingFunction, 'cubic-bezier(0.1, 5, 0.23, 0)');
  assert_equals(computedStyle.transitionTimingFunction, 'cubic-bezier(0.1, 5, 0.23, 0)');
  assert_equals(style.webkitTransitionTimingFunction, 'cubic-bezier(0.1, 5, 0.23, 0)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'cubic-bezier(0.1, 5, 0.23, 0)');

  style.transitionTimingFunction = "cubic-bezier(0.2, 2, 0.2, -0.4)";
  assert_equals(style.transitionTimingFunction, 'cubic-bezier(0.2, 2, 0.2, -0.4)');
  assert_equals(computedStyle.transitionTimingFunction, 'cubic-bezier(0.2, 2, 0.2, -0.4)');
  assert_equals(style.webkitTransitionTimingFunction, 'cubic-bezier(0.2, 2, 0.2, -0.4)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'cubic-bezier(0.2, 2, 0.2, -0.4)');

  style.transitionTimingFunction = "cubic-bezier(0.2, -2, 0.2, 0.4)";
  assert_equals(style.transitionTimingFunction, 'cubic-bezier(0.2, -2, 0.2, 0.4)');
  assert_equals(computedStyle.transitionTimingFunction, 'cubic-bezier(0.2, -2, 0.2, 0.4)');
  assert_equals(style.webkitTransitionTimingFunction, 'cubic-bezier(0.2, -2, 0.2, 0.4)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'cubic-bezier(0.2, -2, 0.2, 0.4)');

  style.transitionTimingFunction = "step-start";
  assert_equals(style.transitionTimingFunction, 'step-start');
  assert_equals(computedStyle.transitionTimingFunction, 'step-start');
  assert_equals(style.webkitTransitionTimingFunction, 'step-start');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'step-start');

  style.transitionTimingFunction = "step-end";
  assert_equals(style.transitionTimingFunction, 'step-end');
  assert_equals(computedStyle.transitionTimingFunction, 'step-end');
  assert_equals(style.webkitTransitionTimingFunction, 'step-end');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'step-end');

  style.transitionTimingFunction = "steps(3)";
  assert_equals(style.transitionTimingFunction, 'steps(3, end)');
  assert_equals(computedStyle.transitionTimingFunction, 'steps(3, end)');
  assert_equals(style.webkitTransitionTimingFunction, 'steps(3, end)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'steps(3, end)');

  style.transitionTimingFunction = "steps(4, end)";
  assert_equals(style.transitionTimingFunction, 'steps(4, end)');
  assert_equals(computedStyle.transitionTimingFunction, 'steps(4, end)');
  assert_equals(style.webkitTransitionTimingFunction, 'steps(4, end)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'steps(4, end)');

  style.transitionTimingFunction = "steps(5, start)";
  assert_equals(style.transitionTimingFunction, 'steps(5, start)');
  assert_equals(computedStyle.transitionTimingFunction, 'steps(5, start)');
  assert_equals(style.webkitTransitionTimingFunction, 'steps(5, start)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'steps(5, start)');

  style.transitionTimingFunction = "frames(5)";
  assert_equals(style.transitionTimingFunction, 'frames(5)');
  assert_equals(computedStyle.transitionTimingFunction, 'frames(5)');
  assert_equals(style.webkitTransitionTimingFunction, 'frames(5)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'frames(5)');

  style.transitionProperty = "opacity, width";

  style.transitionTimingFunction = "ease-in-out, ease-in";
  assert_equals(style.transitionTimingFunction, 'ease-in-out, ease-in');
  assert_equals(computedStyle.transitionTimingFunction, 'ease-in-out, ease-in');
  assert_equals(style.webkitTransitionTimingFunction, 'ease-in-out, ease-in');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease-in-out, ease-in');

  style.transitionTimingFunction = "ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)";
  assert_equals(style.transitionTimingFunction, 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)');
  assert_equals(computedStyle.transitionTimingFunction, 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)');
  assert_equals(style.webkitTransitionTimingFunction, 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)');

  style.transitionTimingFunction = "steps(3, start), ease-in-out";
  assert_equals(style.transitionTimingFunction, 'steps(3, start), ease-in-out');
  assert_equals(computedStyle.transitionTimingFunction, 'steps(3, start), ease-in-out');
  assert_equals(style.webkitTransitionTimingFunction, 'steps(3, start), ease-in-out');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'steps(3, start), ease-in-out');
}, "Valid transition-timing-function values.");

test(() => {
  style.transitionProperty = "";
  style.transitionTimingFunction = "";

  style.transitionTimingFunction = "cubic-bezier(3, 0.1, 4, 1)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "cubic-bezier(1, 0.1, 3, 1)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "cubic-bezier(0.1, 0, 4, 0.4)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "cubic-bezier(3, 0, 4, 0.4)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "cubic-bezier(3, 0, 0.2, 0.4)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "cubic-bezier(-0.2, 0, 0.2, 0.4)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "cubic-bezier(0.2, 2, -0.2, 0.4)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "steps(5, 3)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "steps(-5, start)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "steps(5, start, end)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "step(5)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "frame(5)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "frames()";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "frames(1)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "frames(5, end)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "red";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "cubic-bezier(0.25, 0.1, 0.25)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "cubic-bezier(0.25, 0.1, 0.25, 1, 2)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');

  style.transitionTimingFunction = "cubic-foo(0.25, 0.1, 0.25, 1)";
  assert_equals(style.transitionTimingFunction, '');
  assert_equals(computedStyle.transitionTimingFunction, 'ease');
  assert_equals(style.webkitTransitionTimingFunction, '');
  assert_equals(computedStyle.webkitTransitionTimingFunction, 'ease');
}, "Invalid transition-timing-function values.");

test(() => {
  // Initial test.
  assert_equals(computedStyle.transitionDelay, '0s');
  assert_equals(computedStyle.webkitTransitionDelay, '0s');

  style.transitionDelay = "0s";
  assert_not_equals(Object.keys(style).indexOf('transitionDelay'), -1);
  assert_equals(Object.keys(style).indexOf('webkitTransitionDelay'), -1);
  assert_equals(style.transitionDelay, '0s');
  assert_equals(computedStyle.transitionDelay, '0s');
  assert_equals(style.webkitTransitionDelay, '0s');
  assert_equals(computedStyle.webkitTransitionDelay, '0s');

  style.transitionDelay = "5s";
  assert_equals(style.transitionDelay, '5s');
  assert_equals(computedStyle.transitionDelay, '5s');
  assert_equals(style.webkitTransitionDelay, '5s');
  assert_equals(computedStyle.webkitTransitionDelay, '5s');

  style.transitionDelay = "10ms";
  assert_equals(style.transitionDelay, '10ms');
  assert_equals(computedStyle.transitionDelay, '0.01s');
  assert_equals(style.webkitTransitionDelay, '10ms');
  assert_equals(computedStyle.webkitTransitionDelay, '0.01s');

  style.transitionDelay = "-10ms";
  assert_equals(style.transitionDelay, '-10ms');
  assert_equals(computedStyle.transitionDelay, '-0.01s');
  assert_equals(style.webkitTransitionDelay, '-10ms');
  assert_equals(computedStyle.webkitTransitionDelay, '-0.01s');

  style.transitionProperty = "opacity, width";

  style.transitionDelay = "-10ms, 20s";
  assert_equals(style.transitionDelay, '-10ms, 20s');
  assert_equals(computedStyle.transitionDelay, '-0.01s, 20s');
  assert_equals(style.webkitTransitionDelay, '-10ms, 20s');
  assert_equals(computedStyle.webkitTransitionDelay, '-0.01s, 20s');
}, "Valid transition-delay values.");

test(() => {
  style.transitionProperty = "opacity";
  style.transitionDelay = "";

  style.transitionDelay = "'5ms'";
  assert_equals(style.transitionDelay, '');
  assert_equals(computedStyle.transitionDelay, '0s');
  assert_equals(style.webkitTransitionDelay, '');
  assert_equals(computedStyle.webkitTransitionDelay, '0s');

  style.transitionDelay = "30px";
  assert_equals(style.transitionDelay, '');
  assert_equals(computedStyle.transitionDelay, '0s');
  assert_equals(style.webkitTransitionDelay, '');
  assert_equals(computedStyle.webkitTransitionDelay, '0s');

  style.transitionDelay = "solid";
  assert_equals(style.transitionDelay, '');
  assert_equals(computedStyle.transitionDelay, '0s');
  assert_equals(style.webkitTransitionDelay, '');
  assert_equals(computedStyle.webkitTransitionDelay, '0s');

  style.transitionDelay = "20";
  assert_equals(style.transitionDelay, '');
  assert_equals(computedStyle.transitionDelay, '0s');
  assert_equals(style.webkitTransitionDelay, '');
  assert_equals(computedStyle.webkitTransitionDelay, '0s');

  style.transitionDelay = "20%";
  assert_equals(style.transitionDelay, '');
  assert_equals(computedStyle.transitionDelay, '0s');
  assert_equals(style.webkitTransitionDelay, '');
  assert_equals(computedStyle.webkitTransitionDelay, '0s');

  style.transitionDelay = "0s, 20px";
  assert_equals(style.transitionDelay, '');
  assert_equals(computedStyle.transitionDelay, '0s');
  assert_equals(style.webkitTransitionDelay, '');
  assert_equals(computedStyle.webkitTransitionDelay, '0s');
}, "Invalid transition-delay values.");

test(() => {
  style.transition = "";
  // Initial test.
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "none";
  assert_not_equals(Object.keys(style).indexOf('transition'), -1);
  assert_equals(Object.keys(style).indexOf('webkitTransition'), -1);
  assert_equals(style.transition, 'none');
  assert_equals(computedStyle.transition, 'none 0s ease 0s');
  assert_equals(style.webkitTransition, 'none');
  assert_equals(computedStyle.webkitTransition, 'none 0s ease 0s');
  assert_true(checkTransitionShorthandValue());

  style.transition = "none 20s";
  assert_equals(style.transition, 'none 20s');
  assert_equals(computedStyle.transition, 'none 20s ease 0s');
  assert_equals(style.webkitTransition, 'none 20s');
  assert_equals(computedStyle.webkitTransition, 'none 20s ease 0s');
  assert_true(checkTransitionShorthandValue());

  style.transition = "ease-in none 20s";
  assert_equals(style.transition, 'none 20s ease-in');
  assert_equals(computedStyle.transition, 'none 20s ease-in 0s');
  assert_equals(style.webkitTransition, 'none 20s ease-in');
  assert_equals(computedStyle.webkitTransition, 'none 20s ease-in 0s');
  assert_true(checkTransitionShorthandValue());

  style.transition = "ease-in opacity 20s";
  assert_equals(style.transition, 'opacity 20s ease-in');
  assert_equals(computedStyle.transition, 'opacity 20s ease-in 0s');
  assert_equals(style.webkitTransition, 'opacity 20s ease-in');
  assert_equals(computedStyle.webkitTransition, 'opacity 20s ease-in 0s');
  assert_true(checkTransitionShorthandValue());

  style.transition = "ease-in opacity 20s 10s";
  assert_equals(style.transition, 'opacity 20s ease-in 10s');
  assert_equals(computedStyle.transition, 'opacity 20s ease-in 10s');
  assert_equals(style.webkitTransition, 'opacity 20s ease-in 10s');
  assert_equals(computedStyle.webkitTransition, 'opacity 20s ease-in 10s');
  // Let's double check here the delay and duration. As stated in the spec the first value parsed
  // is assigned to the duration.
  assert_equals(computedStyle.transitionDuration, '20s');
  assert_equals(computedStyle.webkitTransitionDuration, '20s');
  assert_equals(computedStyle.transitionDelay, '10s');
  assert_equals(computedStyle.webkitTransitionDelay, '10s');
  assert_true(checkTransitionShorthandValue());

  style.transition = "20s ease-in opacity 10s";
  assert_equals(style.transition, 'opacity 20s ease-in 10s');
  assert_equals(computedStyle.transition, 'opacity 20s ease-in 10s');
  assert_equals(style.webkitTransition, 'opacity 20s ease-in 10s');
  assert_equals(computedStyle.webkitTransition, 'opacity 20s ease-in 10s');
  assert_equals(computedStyle.transitionDuration, '20s');
  assert_equals(computedStyle.webkitTransitionDuration, '20s');
  assert_equals(computedStyle.transitionDelay, '10s');
  assert_equals(computedStyle.webkitTransitionDelay, '10s');
  assert_true(checkTransitionShorthandValue());

  style.transition = "ease-in opacity 20s 10s, cubic-bezier(0.32, 0, 1, 1) 10s width 20s";
  assert_equals(style.transition, 'opacity 20s ease-in 10s, width 10s cubic-bezier(0.32, 0, 1, 1) 20s');
  assert_equals(computedStyle.transition, 'opacity 20s ease-in 10s, width 10s cubic-bezier(0.32, 0, 1, 1) 20s');
  assert_equals(style.webkitTransition, 'opacity 20s ease-in 10s, width 10s cubic-bezier(0.32, 0, 1, 1) 20s');
  assert_equals(computedStyle.webkitTransition, 'opacity 20s ease-in 10s, width 10s cubic-bezier(0.32, 0, 1, 1) 20s');
  // Let's double check here the delay and duration. As stated in the spec the first value parsed
  // is assigned to the duration.
  assert_equals(computedStyle.transitionDuration, '20s, 10s');
  assert_equals(computedStyle.webkitTransitionDuration, '20s, 10s');
  assert_equals(computedStyle.transitionDelay, '10s, 20s');
  assert_equals(computedStyle.webkitTransitionDelay, '10s, 20s');
  assert_true(checkTransitionShorthandValue());

  style.transition = "all, all";
  assert_equals(style.transition, 'all, all');
  assert_equals(computedStyle.transition, 'all 0s ease 0s, all 0s ease 0s');
  assert_equals(style.webkitTransition, 'all, all');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s, all 0s ease 0s');

  style.transition = "all 20s 10s ease-in, all ease-out 20s";
  assert_equals(style.transition, 'all 20s ease-in 10s, all 20s ease-out');
  assert_equals(computedStyle.transition, 'all 20s ease-in 10s, all 20s ease-out 0s');
  assert_equals(style.webkitTransition, 'all 20s ease-in 10s, all 20s ease-out');
  assert_equals(computedStyle.webkitTransition, 'all 20s ease-in 10s, all 20s ease-out 0s');

  style.transition = "ease-in opacity 20s 10s, all ease-out 20s";
  assert_equals(style.transition, 'opacity 20s ease-in 10s, all 20s ease-out');
  assert_equals(computedStyle.transition, 'opacity 20s ease-in 10s, all 20s ease-out 0s');
  assert_equals(style.webkitTransition, 'opacity 20s ease-in 10s, all 20s ease-out');
  assert_equals(computedStyle.webkitTransition, 'opacity 20s ease-in 10s, all 20s ease-out 0s');
  assert_true(checkTransitionShorthandValue());

  style.transition = " all ease-out 20s, ease-in opacity 20s 10s";
  assert_equals(style.transition, 'all 20s ease-out, opacity 20s ease-in 10s');
  assert_equals(computedStyle.transition, 'all 20s ease-out 0s, opacity 20s ease-in 10s');
  assert_equals(style.webkitTransition, 'all 20s ease-out, opacity 20s ease-in 10s');
  assert_equals(computedStyle.webkitTransition, 'all 20s ease-out 0s, opacity 20s ease-in 10s');
  assert_true(checkTransitionShorthandValue());
}, "Valid transition shorthand values.");

test(() => {
  style.transition = "";

  style.transition = "20";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "20, 20";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "all 30s width ease-in";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "all 30s ease-in 20px";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "all 30s ease-in 20s, 20px";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "all 30s ease-in 20s, step-start(2)";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "ease-in opacity 20s 10s, none";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "none, ease-in opacity 20s 10s, none";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "none, ease-in opacity 20s 10s";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "ease-in opacity 20s 10s, all 20s ease-out 0s, none";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "ease-in opacity 20s 10s, ease-in width 20s 10s, none";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');

  style.transition = "none, none";
  assert_equals(style.transition, '');
  assert_equals(computedStyle.transition, 'all 0s ease 0s');
  assert_equals(style.webkitTransition, '');
  assert_equals(computedStyle.webkitTransition, 'all 0s ease 0s');
}, "Invalid transition shorthand values.");

test(() => {
  style.transition = "ease-in ease-out";
  assert_equals(style.transitionProperty, 'ease-out');
  assert_equals(style.transitionTimingFunction, 'ease-in');

  style.transition = "20s 10s";
  assert_equals(style.transitionDelay, '10s');
  assert_equals(style.transitionDuration, '20s');

  style.transition = "-10s 10s";
  assert_equals(style.transitionDelay, '-10s');
  assert_equals(style.transitionDuration, '10s');

  style.transition = "ease";
  assert_equals(style.transitionTimingFunction, 'ease');
  assert_equals(style.transitionProperty, 'all'); // Fails: crbug.com/772772
}, "Ambiguous shorthand values.");
</script>
</body>
</html>
